requirements.txt
Flask
Flask-SQLAlchemy
pyotp
qrcode
pillow
PyJWT
pymysql
python-dotenv
